EntitySpawner - Fix burst errors
