+++ Scripts/Runtime/Entities/Lifecycle/Spawner/BurstEntitySpawner.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 0199571a357f49eba991ebd5634d6c1e
+timeCreated: 1681937654
